backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
langchain>=0.3.1
langgraph>=0.2.34
langchain-google-genai>=2.0.0
langchain-ollama>=0.2.0
python-dotenv>=1.0.1
pydantic>=2.9.2
websockets>=13.1
black
isort
chromadb>=0.4.0
pypdf>=4.0.0
tiktoken>=0.7.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.2.0
langchain-community>=0.2.0
langgraph-checkpoint-sqlite>=1.0.0
